no-extraneous-dependencies includes only paths that can be resolved, as only resolved paths can be reliably tested if they live in a `node_modules` folder.
Up until now, this rule assumed, that every package that lives in a `node_modules` folder
or *cannot be resolved* is an external package.
This happens because of [that line](https://github.com/benmosher/eslint-plugin-import/blob/master/src/core/importType.js#L32).

If project uses aliasing, or transforms imports, this assumption is wrong,
as they may not be resolvable by import plugin, but also, they should not be defined in package.json.

Changing `isExternalPath` will make any ordering lints unstable
(order may change depending if you ran `npm install` or not).

Using `import/ignore` setting does not make sense too, as it's used to ignore imports unknown to the plugin,
like .css files. But still, you may import a css file from a dependency, and this rule should check that.